+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - 2025-04-06
+
+### Added
+- Changes in .github/workflow/python-package.yml
+- Changes in CONTRIBUTION.md
+- Changes in MANIFEST.in
+- Changes in Makefile
+- Changes in pytest.ini
+- Changes in src/infrash_embedded/analyzer.py
+- Changes in src/infrash_embedded/cli.py
+- Changes in tests/test_analyzer.py
+- Changes in tests/test_cli.py
+- Changes in tox.ini
+
+### Removed
+- Changes in src/infrash_embedded/__init__.py.bak
+- Changes in src/infrash_embedded/_version.py.bak
+
 ## [0.1.1] - 2025-04-06
 
 ### Added